+++ Paper.bib
@@ -6,7 +6,7 @@
   url = {https://CRAN.R-project.org/package=OptimClassifier},
 }
 
- @Book{Fox:2017,
+@Book{Fox:2017,
     title = {Using the {R Commander}: A Point-and-Click Interface for {R}},
     author = {John Fox},
     year = {2017},
@@ -22,9 +22,10 @@
   journal = {R News},
   volume = 7,
   issue = 3,
-  page = {46--52}
+  page = {46-52},
   url = {http://CRAN.R-project.org/doc/Rnews/}
   }
+
 @Manual{R:2018,
     title = {R: A Language and Environment for Statistical Computing},
     author = {{R Core Team}},
