### 2018-04-23 cc58866
Remove double braquets
