+++ paper.md
@@ -32,11 +32,19 @@ bibliography: Paper.bib
 
 The R Commander (``Rcmdr``) package [@Fox:2017] offers a point-and-click graphical user interface (GUI), despite to was originally conceived as a basic-statistics graphical user interface to R [@R:2018].
 
-Nowadays, the  Rcmdr includes facilities for fitting, checking, and displaying statistical models in general and plugins for specific areas. This is because its menu and dialog structure were designed to be extensible, extending the Rcmdr became much more convenient with the introduction of the plug-in packages, described in @Fox:2007,  and as we write this, 56 ``Rcmdr`` plug-ins are available on CRAN and 17 ``Rcmdr`` plug-ins are available on Github.
+Nowadays, the  Rcmdr includes facilities for fitting, checking, and displaying several statistical models. This is because its menu and dialog structure were designed to be extensible, extending the Rcmdr became much more convenient with the introduction of the plug-in packages, described in @Fox:2007,  and as we write this, 56 ``Rcmdr`` plug-ins are available on CRAN and 17 ``Rcmdr`` plug-ins are available on Github.
 
+Each of statistical methods has parameters inputs that we can optimize, so that return output will improve respect default options. OptimClassifier package (@Perez:2018, @R:2018) is useful in optimization and  modelling to obtain the best parameters for a classification model.
+The main goal of ``RcmdrPlugin.OptimClassifier`` is providing a point-and-click graphical user interface (GUI) to ``OptimClassifier package``.``RcmdrPlugin.OptimClassifier`` currently already provides a whole toolbox of easy to use training models:
+ - Linear Models (LM)
+ - Generalized Linear Models (GLM)
+ - Linear Mixed Models (LMM)
+ - Discriminant Analysis (DA)
+ - Classification And Regression Trees (CART)
+ - Neural Networks (NN)
+ - Support Vector Machines (SVM) 
 
-Each of stadistical methods has parameters inputs that we can optimize, so that return output will improve respect default options. The main goal of ``RcmdrPlugin.OptimClassifier`` is providing a point-and-click graphical user interface (GUI) to OptimClassifier package (@Perez:2018, @R:2018 ). OptimClassifier package is an R package that collects most popular methodologies and compares different configuration parameters to find the best choice.
 ![Some example of Optimum Support Vector Machines dialog](SVMMenu.png)
- ``RcmdrPlugin.OptimClassifier`` allows to fit the most important statistical methodologies included in ``OptimClassifier`` with simple dialogs. The source code for ``RcmdrPlugin.OptimClassifier`` has been archived to Zenodo with the linked DOI: @zenodo and Github
+The source code for ``RcmdrPlugin.OptimClassifier`` has been archived to Zenodo with the linked DOI: @zenodo and Github
 
 # References
