### 2018-04-23 1ec6af4
Bib fixes
