+++ Paper.bib
@@ -1,6 +1,6 @@
 @Manual{Perez:2018,
   title = {OptimClassifier: Create the Best Train for Classification Models},
-  author = {Agustín Pérez-Martín and Agustín Pérez-Torregrosa and Marta Vaca-Lamata and Antonio José Verdú-Jover},
+  author = {Pérez-Martín, Agustín and Pérez-Torregrosa, Agustín and Vaca-Lamata, Marta and  Verdú-Jover, Antonio José},
   year = {2018},
   note = {R package version 0.1.4},
   url = {https://CRAN.R-project.org/package=OptimClassifier},
@@ -8,7 +8,7 @@
 
 @Book{Fox:2017,
     title = {Using the {R Commander}: A Point-and-Click Interface for {R}},
-    author = {John Fox},
+    author = {Fox, John},
     year = {2017},
     publisher = {Chapman and Hall/CRC Press},
     address = {Boca Raton {FL}},
@@ -17,7 +17,7 @@
 
 @article{Fox:2007,
   title = {Extending the {R Commander} by {‘Plug-In’ Packages}},
-  author = {John Fox},
+  author = {Fox, John},
   year = {2007},
   journal = {R News},
   volume = 7,
@@ -44,9 +44,9 @@
        Discriminant Analysis (DA), Classification And Regression Trees (CART), Neural Networks (NN)
        and Support Vector Machines (SVM).},
     publisher={Zenodo},
-    author={Agustin Pérez-Martín and
-          Agustín Pérez-Torregrosa and
-          Marta Vaca-Lamata and
-          Antonio José Verdú-Jover},
+    author={Pérez-Martín, Agustín and
+          Pérez-Torregrosa, Agustín and
+          Vaca-Lamata, Marta and
+          Verdú-Jover, Antonio José},
     year={2018}
     }
