### 2018-04-27 #697 @economistgame
I suggest @wrathematics and @masalmon to review the paper
