### 2018-05-03 c6a8fa4
Changes in TravisCI
