+++ .travis.yml
@@ -1,5 +1,16 @@
-# R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
+ #R for travis: see documentation at https://docs.travis-ci.com/user/languages/r
 
 language: R
+os:
+- linux
+- osx
 sudo: false
 cache: packages
+
+r_packages:
+  - covr
+  - car
+  - carData
+
+after_success:
+- Rscript -e 'library(covr); codecov()'
