### 2018-05-03 a11a878
Changes in Travis CI
