### 2018-05-07 #697 @karthik
@economistgame Thanks for the suggestions.

@andrewheiss and/or @briatte, would either of you be able to review?
