### 2018-05-07 #697 @karthik
Thanks @andrewheiss! Once I start the review you'll be tagged in another issue with a checklist and link to guidelines (http://joss.theoj.org/about#reviewer_guidelines). I'd be happy to answer any questions you may have about the process (email in profile).
