### 2018-05-07 #727 @whedon
[REVIEW]: RcmdrPlugin.OptimClassifier: Extending the R Commander Interface to create the best train for classification models
**Submitting author:** @economistgame (<a href="http://orcid.org/0000-0001-5658-4795">Agustín Pérez-Torregrosa</a>)
**Repository:** <a href="https://github.com/economistgame/RcmdrPlugin.OptimClassifier" target ="_blank">https://github.com/economistgame/RcmdrPlugin.OptimClassifier</a>
**Version:** v0.1.2

@@andrewheiss, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:
