### 2018-05-16 economistgame/RcmdrPlugin.OptimClassifier#2 @andrewheiss
Optim.LMM possibly conflicting with other R Commander packages
When running `Optim.LMM()` after loading **RcmdrPlugin.OptimClassifier**, it occasionally gives errors. I'm guessing there's some sort of package namespace conflict behind the scenes, but I can't uncover it. The model works fine when run manually with just the **OptimClassifier** library.

Here's a reproducible example using **OptimClassifier** 0.1.4 and **RcmdrPlugin.OptimClassifier** 0.1.2:

    # This works
    library(OptimClassifier)
    modelFit <- Optim.LMM("Y", AustralianCredit, p = 0.7, seed = 2018)
    summary(modelFit)

    # Start a new session of R. This doesn't work
    library(RcmdrPlugin.OptimClassifier)
    modelFit <- Optim.LMM("Y", AustralianCredit, p = 0.7, seed = 2018)
    #> Error in eval(predvars, data, env) : object 'X4' not found

Review issue: https://github.com/openjournals/joss-reviews/issues/727
