### 2018-05-16 economistgame/RcmdrPlugin.OptimClassifier#3 @andrewheiss
Optim.LM dialog doesn't create a matrix-based formula
This may be an issue with **OptimClassifier** 0.1.4, but it also has a Rcmdr component. `OptimClassifier::Optim.LM()` says that it takes a formula as an argument, but when you pass a formula, it complains that the formula object is not a matrix:

    Optim.LM(Y ~ X1 + X2 + X3 + X4 + X5 + X6, data = AustralianCredit, p = 0.7,
             criteria = 'success_rate', seed = 2018)
    #> Error in model.frame.default(formula = ModelsTested$train_formula[[i]],  :
    #> object is not a matrix

As a result, it's not currently possible to select a few variables using R Commander. For instance, this:

…results in this:

![image](https://user-images.githubusercontent.com/73663/40096215-25104dd0-588d-11e8-8771-3d8b170426f0.png)

`Optim.LM()` *does* work if the formula uses a `.` for the RHS, but there's no way to generate that with R Commander—selecting all the variables just concatenates them into a long formula:

    # This works, but can't be generated with R Commander
    Optim.LM(Y~., AustralianCredit, p = 0.7, seed = 2018)

    #          AustralianCredit, p = 0.7, seed = 2018)

Review issue: https://github.com/openjournals/joss-reviews/issues/727
