### 2018-05-16 economistgame/RcmdrPlugin.OptimClassifier#4 @andrewheiss
Unable to create diagnostic plots for DA models
When running "Models" > "Graphs" > "Basic diagnostic plots" on a discriminant analysis model (or when just running `plot()`, plotting fails.

    OptimDA2 <- Optim.DA(Y~X1+X2+X3+X4+X5+X6+X7+X8+X9+X10+X11+X12+X13+X14,
                         data=AustralianCredit, p=0.7, criteria='rmse',seed=2018)
    #> Error in rep("darkgrey", nrow(Files) - 2) : invalid 'times' argument

Review issue: https://github.com/openjournals/joss-reviews/issues/727
