### 2018-05-16 economistgame/RcmdrPlugin.OptimClassifier#5 @andrewheiss
Rplots.pdf unncessary
There's currently an empty/corrupt `Rplots.pdf` in the root of the repository that doesn't need to be there and can be removed from the repo.

Review issue: https://github.com/openjournals/joss-reviews/issues/727
