### 2018-05-16 b28af23
Some editorial suggestions for the JOSS paper
