+++ paper.md
@@ -32,10 +32,11 @@ bibliography: Paper.bib
 
 The R Commander (``Rcmdr``) package [@Fox:2017] offers a point-and-click graphical user interface (GUI), despite to was originally conceived as a basic-statistics graphical user interface to R [@R:2018].
 
-Nowadays, the  Rcmdr includes facilities for fitting, checking, and displaying several statistical models. This is because its menu and dialog structure were designed to be extensible, extending the Rcmdr became much more convenient with the introduction of the plug-in packages, described in @Fox:2007,  and as we write this, 56 ``Rcmdr`` plug-ins are available on CRAN and 17 ``Rcmdr`` plug-ins are available on Github.
+Nowadays, ``Rcmdr`` includes facilities for fitting, checking, and displaying many types of statistical models. With the introduction of plug-in packages, described in @Fox:2007, it is convenient to extend ``Rcmdr`` to include a range of additional features. As we write this, 56 ``Rcmdr`` plug-ins are available on CRAN and 17 ``Rcmdr`` plug-ins are available on GitHub.
 
-Each of statistical methods has parameters inputs that we can optimize, so that return output will improve respect default options. OptimClassifier package (@Perez:2018, @R:2018) is useful in optimization and  modelling to obtain the best parameters for a classification model.
-The main goal of ``RcmdrPlugin.OptimClassifier`` is providing a point-and-click graphical user interface (GUI) to ``OptimClassifier`` package. ``RcmdrPlugin.OptimClassifier`` currently already provides a whole toolbox of easy to use training models:
+The main goal of ``RcmdrPlugin.OptimClassifier`` is to provide a point-and-click GUI for the ``OptimClassifier`` package [@Perez:2018; @R:2018]. Each of the available statistical methods has inputs that we can optimize so that return output will improve with respect to default options. ``OptimClassifier`` is useful for obtaining the best parameters for a classification model for optimization and modelling.
+
+``RcmdrPlugin.OptimClassifier`` currently provides dialogs for a toolbox of easy to use training models:
 
  - Linear Models (LM)
  - Generalized Linear Models (GLM)
@@ -46,6 +47,7 @@ The main goal of ``RcmdrPlugin.OptimClassifier`` is providing a point-and-click 
  - Support Vector Machines (SVM) 
 
 ![Some example of Optimum Support Vector Machines dialog](SVMMenu.png)
+
 The source code for ``RcmdrPlugin.OptimClassifier`` has been archived to Zenodo with the linked DOI: @zenodo and Github
 
 # References
