### 2018-05-16 c60d7b8
Small editorial suggestions
