### 2018-05-16 #727 @andrewheiss
**RcmdrPlugin.OptimClassifier** is a helpful new package that provides a menu-based interface for newer machine learning classification models. It was easy to install and use, and I only came across a few issues, noted below.

The paper has a good, succinct justification for why this library is needed, but the README doesn't. It might be helpful to include some of that justification there (noting that R Commander provides a graphic interface for building models in R, etc.). Additionally, the paper has a list of models that the plugin allows users to build; the README doesn't and probably should.

The paper says that the menus provide 7 different models, but the actual R Commander menus only list 6—classification and regression trees (CART) is not an option.

I've opened a few issues that I ran into when testing and fitting models. I'm not sure if all the problems I ran into are specific to **RcmdrPlugin.OptimClassifier** or just **OptimClassifier**, but they definitely affect the use of  **RcmdrPlugin.OptimClassifier** so they should probably be fixed.

It might be helpful to add some tests to ensure that the `OptimClassifier::Optim*()` functions continue to work after loading **RcmdrPlugin.OptimClassifier**. In its current form, there are no tests. Part of this is probably because R Commander plugins are harder to write tests for, since they're based on a GUI (my own R Commander plugin package doesn't have any tests 😬—I should fix that someday). However, in this situation, with the possible conflicting namespace issues noted in [https://github.com/economistgame/RcmdrPlugin.OptimClassifier/issues/2](https://github.com/economistgame/RcmdrPlugin.OptimClassifier/issues/2), some test coverage could catch these problems.

It also might be helpful to include an explanation or screenshot in the README or the JOSS article indicating where to find the new **OptimClassifier** menu options (similar to [this](https://github.com/andrewheiss/RcmdrPlugin.MPAStats)). After loading data, the "Models" menu is still mostly greyed out—users have to go to "Fit models" \> "Optimum XXX". Part of this confusion is due to how R Commander works—it only shows stuff in the "Models" menu after a model has been fit, and models can only be fit with "Statistics" \> "Fit models". Longtime users of R Commander will know this, but new users (the target audience of R Commander) might struggle with where to find the new menu items.

Explaining what else the package offers could be helpful. As it stands now, it's most obvious that it adds six menu items under "Statistics" \> "Fit models". Each of those model types have a few additional options under the "Models" menu—users can summarize them and create diagnostic plots for them (all other model options are greyed out). It's not immediately clear that users can do this with the newly-fit models though.

It would be helpful to have some community guidelines in the package too, such as a CONTRIBUTING.md file (perhaps modeled after something like [this](https://github.com/ropensci/nomisr/blob/master/CONTRIBUTING.md) or [this](https://github.com/tidyverse/ggplot2/blob/master/CONTRIBUTING.md)) and a CONDUCT.md file (like [this](https://github.com/ropensci/qualtRics/blob/master/CONDUCT.md) or [this](https://github.com/tidyverse/readr/blob/master/CONDUCT.md))

One reference in the BibTeX file needs a DOI: *Using the R Commander*: 10.1201/9781315380537. Perez:2018 doesn't have a DOI yet (but will?), and Fox:2007 and R:2018 don't have one assigned to them as far as I can tell, so they're okay.

Some of the English in the README and the paper needs to be revised and edited. I have suggested a number of edits in [https://github.com/economistgame/RcmdrPlugin.OptimClassifier/pull/6](https://github.com/economistgame/RcmdrPlugin.OptimClassifier/pull/6) and [https://github.com/economistgame/RcmdrPlugin.OptimClassifier/pull/7](https://github.com/economistgame/RcmdrPlugin.OptimClassifier/pull/7). In addition to those suggestions, the title doesn't need to have "Interface," "Classification," and "Models" capitalized.

Phew. That's all I can see. This is great!
