### 2018-05-18 economistgame/RcmdrPlugin.OptimClassifier#5 @economistgame
Okey solved
