### 2019-04-10 #727 @arfon
It's now been more than six months since we've heard from @economistgame so I'm going to assume they are no-longer interested in pursuing this publication with JOSS.

@andrewheiss @karthik thanks for your help with this paper. Sorry this didn't make it completely through review.
